requirements.txt
Flask==3.1.2
Flask_Limiter==3.5.0
pyserial==3.5
python-dotenv==1.2.1
Requests==2.32.5
